+++ packages/minapp-vscode/src/extension.ts
@@ -5,31 +5,36 @@ Author Mora <qiuzhongleiabc@126.com> (https://github.com/qiu8310)
 
 import {ExtensionContext, languages} from 'vscode'
 
-import WxmlAutoCompletion from './plugin/WxmlAutoCompletion'
-import WxmlHoverProvider from './plugin/WxmlHoverProvider'
+import HoverProvider from './plugin/HoverProvider'
 import WxmlFormatter from './plugin/WxmlFormatter'
+
+import WxmlAutoCompletion from './plugin/WxmlAutoCompletion'
+import PugAutoCompletion from './plugin/PugAutoCompletion'
+import VueAutoCompletion from './plugin/VueAutoCompletion'
+
 import {config, configActivate, configDeactivate} from './plugin/lib/config'
 
 export function activate(context: ExtensionContext) {
   configActivate()
-
-  let selector = ['wxml', 'vue']
   let formatter = new WxmlFormatter()
-  let autoCompletion = new WxmlAutoCompletion(config)
-  let hoverProvider = new WxmlHoverProvider(config)
+  let autoCompletionWxml = new WxmlAutoCompletion(config)
+  let hoverProvider = new HoverProvider(config)
+  let autoCompletionPug = new PugAutoCompletion(config)
+  let autoCompletionVue = new VueAutoCompletion(autoCompletionPug, autoCompletionWxml)
 
   // console.log('minapp-vscode is active!')
   context.subscriptions.push(
-    languages.registerHoverProvider(selector, hoverProvider),
-    languages.registerCompletionItemProvider(selector, autoCompletion, '<', ' ', ':', '@', '.', '-'),
-    // languages.registerOnTypeFormattingEditProvider('wxml', formatter, '\n', '\t', 'a'),
-    languages.registerDocumentFormattingEditProvider('wxml', formatter),
-    languages.registerDocumentRangeFormattingEditProvider('wxml', formatter)
-  )
+    languages.registerHoverProvider(['wxml', 'wxml-pug', 'vue'], hoverProvider),
 
-  return {
-    autoCompletion, hoverProvider
-  }
+    // 格式化
+    languages.registerDocumentFormattingEditProvider('wxml', formatter),
+    languages.registerDocumentRangeFormattingEditProvider('wxml', formatter),
+
+    // 自动补全
+    languages.registerCompletionItemProvider('wxml', autoCompletionWxml, '<', ' ', ':', '@', '.', '-'),
+    languages.registerCompletionItemProvider('wxml-pug', autoCompletionPug, '\n', ' ', ':', '@', '.', '-'),
+    languages.registerCompletionItemProvider('vue', autoCompletionVue, '<', ' ', ':', '@', '.', '-', '\n')
+  )
 }
 
 export function deactivate() {