+++ packages/minapp-vscode/src/plugin/HoverProvider.ts
@@ -5,16 +5,26 @@ Author Mora <qiuzhongleiabc@126.com> (https://github.com/qiu8310)
 
 import {HoverProvider, TextDocument, Position, CancellationToken, Hover, MarkdownString} from 'vscode'
 import { hoverComponentAttrMarkdown, hoverComponentMarkdown } from '@minapp/common'
-import {getTagAtPosition, Tag} from './getTagAtPosition'
+import {Tag} from './lib/getTagAtPosition'
+import {getTagAtPosition as getWxmlTag} from './lib/getTagAtPositionForWxml'
+import {getTagAtPosition as getPugTag} from './lib/getTagAtPositionForPug'
 import {Config} from './lib/config'
-import {getLanguage, getCustomOptions} from './lib/helper'
+import {getLanguage, getCustomOptions, getLangForVue} from './lib/helper'
 
 export default class implements HoverProvider {
   constructor(public config: Config) {}
 
   async provideHover(document: TextDocument, position: Position, token: CancellationToken) {
+    let lang = document.languageId
+    if (lang === 'vue') {
+      lang = getLangForVue(document, position) as string
+      if (!lang) return null
+    }
+
     let language = getLanguage(document, position)
     if (!language) return null
+
+    let getTagAtPosition = /pug/.test(lang) ? getPugTag : getWxmlTag
     let tag = getTagAtPosition(document, position) as Tag
     if (!tag) return null